backend/requirements.txt
fastapi
uvicorn
pydantic
langchain-core
langchain-groq

pydantic
python-dotenv
langchain
langchain-groq
